Rename some files and Match spawn_players.c (#226)
* name some code files

* Match spawn_players func

* random decomp

* name more files

+++ include/variables.h
@@ -140,7 +140,7 @@ extern s16  D_801651D0[4][8];
 
 extern f32  D_80165210[];
 extern f32  D_80165230[];
-extern s32  D_8016524C;
+extern f32  D_8016524C;
 extern s16  D_80165270[8];
 extern f32  D_80165280[];
 extern f32  D_801652A0[];
